requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
psycopg2-binary  # if using Postgres
stripe
celery
redis  # for Celery broker
phonenumbers
Pillow  # for image handling
django-filter  # for filtering in DRF
django-extensions  # useful development tools